+++ src/Requests/Products/GetProductTagTypesRequest.php
@@ -1,0 +1,7 @@
+<?php namespace Ordercloud\Requests\Products;
+
+use Ordercloud\Support\CommandBus\Command;
+
+class GetProductTagTypesRequest implements Command
+{
+}