+++ src/Requests/Products/Handlers/DisableProductTagRequestHandler.php
@@ -1,0 +1,15 @@
+<?php namespace Ordercloud\Requests\Products\Handlers;
+
+use Ordercloud\Requests\Handlers\AbstractPutRequestHandler;
+use Ordercloud\Requests\Products\EnableProductTagRequest;
+
+class DisableProductTagRequestHandler extends AbstractPutRequestHandler
+{
+    /**
+     * @param EnableProductTagRequest $request
+     */
+    protected function configure($request)
+    {
+        $this->setUrl('/resource/product/tag/%d/change/disable', $request->getTagId());
+    }
+}