+++ src/Services/ProductService.php
@@ -5,11 +5,13 @@ use Ordercloud\Entities\Products\ProductCollection;
 use Ordercloud\Entities\Products\ProductTag;
 use Ordercloud\Requests\Products\Criteria\ProductCriteria;
 use Ordercloud\Requests\Products\Criteria\ProductsByConnectionCriteria;
+use Ordercloud\Requests\Products\Criteria\TagCriteria;
 use Ordercloud\Requests\Products\FindProductsRequest;
 use Ordercloud\Requests\Products\GetProductRequest;
 use Ordercloud\Requests\Products\GetProductsByConnectionRequest;
 use Ordercloud\Requests\Products\GetProductTagRequest;
 use Ordercloud\Requests\Products\GetProductTagsForOrganisationByTypeNameRequest;
+use Ordercloud\Requests\Products\GetProductTagsRequest;
 
 class ProductService extends OrdercloudService
 {
@@ -40,28 +42,33 @@ class ProductService extends OrdercloudService
     }
 
     /**
+     * @deprecated This has been replaced by `ProductTagService::getTags()`
+     *
      * @param $organisationId
-     * @param $tagName
+     * @param $typeName
      *
      * @return array|ProductTag[]
      */
-    public function getProductTags($organisationId, $tagName)
+    public function getProductTags($organisationId, $typeName)
     {
-        return $this->request(
-            new GetProductTagsForOrganisationByTypeNameRequest($organisationId, $tagName)
-        );
+        $criteria = TagCriteria::create()
+            ->setOrganisationId($organisationId)
+            ->setTypeName($typeName)
+            ->setPageSize(-1);
+
+        return $this->request(new GetProductTagsRequest($criteria));
     }
 
     /**
+     * @deprecated This has been replaced by `ProductTagService::getTag()`
+     *
      * @param $tagId
      *
      * @return ProductTag
      */
     public function getProductTag($tagId)
     {
-        return $this->request(
-            new GetProductTagRequest($tagId)
-        );
+        return $this->request(new GetProductTagRequest($tagId));
     }
 
     /**
@@ -73,7 +80,7 @@ class ProductService extends OrdercloudService
     public function getProductsByConnection($connectionId, ProductsByConnectionCriteria $criteria = null)
     {
         $criteria = $criteria ?: ProductsByConnectionCriteria::create();
-        
+
         return $this->request(
             new GetProductsByConnectionRequest($connectionId, $criteria)
         );